fix(storage): enforce exact-generation provider cleanup evidence
Concurrently attempt every retained provider stop, emit canonical evidence once, and fail closed for unowned retained generations. Keep cache completion evidence in its cleanup scope and split oversized stop/cache regression modules.

Closes #1792

Refs #1710
