+++ changes/1792.fixed.md
@@ -1,7 +1,11 @@
-Process projection cache cleanup now records provider stops in canonical
-component order with each exact mount ID and its observed STOP acknowledgement,
-then listener, lease-resource, cleanup-ledger, projection-root, cache-removal,
-and completion evidence in one bounded execution. The first failure terminates
-the evidence stream. Test-owned join handles remain owned through timeout and
-body-panic unwinds, are aborted only when necessary, and reach terminal join
-settlement before secondary cleanup failures are reported.
+Process projection cache cleanup concurrently attempts every retained Branch,
+OwnerHome, and FleetShared provider stop, then emits their exact outcomes in
+canonical component order and terminates emitted success evidence at the first
+canonical failure. Exact-generation retained settlement remains distinct with
+`acknowledged: false`; an unowned provider generation fails closed instead of
+probing ambient or replacement endpoint state. The complete evidence sequence
+stays in one scope, with listener settlement emitted once after binding and
+provider stops and last-close cache evidence retained through completion. The
+oversized cache regression module uses a focused evidence-test sibling. Test
+join handles retain their accepted timeout, abort, join, and primary-panic
+mechanics.
