+++ crates/astrid-kernel/src/storage_mount/process_broker/lease_atomicity_tests/cache_invalidation_tests.rs
@@ -34,6 +34,9 @@ use crate::storage_mount::{MountOwnerScope, issue_lease, revoke_lease};
 #[path = "root_retry_tests.rs"]
 mod root_retry_tests;
 
+#[path = "cache_invalidation_evidence_tests.rs"]
+mod evidence_tests;
+
 fn provider_test_executable(
     directory: &std::path::Path,
     name: &str,
@@ -135,6 +138,8 @@ macro_rules! owned_finishers {
     };
 }
 
+pub(crate) use owned_finishers;
+
 async fn bounded_until(phase: &'static str, mut predicate: impl FnMut() -> bool) {
     bounded_phase(phase, async {
         while !predicate() {
@@ -155,6 +160,8 @@ macro_rules! provider_fixture {
         (temporary, kernel, caller, broker)
     }};
 }
+
+pub(crate) use provider_fixture;
 
 async fn join_mount(
     task: &Arc<OwnedTask<Result<astrid_capsule::context::ProcessStorageMount, String>>>,
@@ -416,377 +423,6 @@ async fn fresh_admission_has_exactly_one_guard_reference() {
     assert_eq!(remounted_projection.refs.load(Ordering::Acquire), 0);
     assert!(kernel.storage_mounts.is_empty());
     assert!(broker.projections.lock().await.is_empty());
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn revoked_component_invalidates_cached_exact_set() {
-    let (_temporary, kernel, caller, broker) =
-        provider_fixture!("revoked_component_invalidates_cached_exact_set");
-    let stale = successful_fleet_mount_for_fresh_execution(&kernel, &caller, &broker).await;
-    let revoked_mount_id = stale.projection.component_mount_ids[1];
-
-    revoke_lease(
-        &kernel,
-        &caller,
-        MountOwnerScope::CallerOnly,
-        revoked_mount_id,
-    )
-    .await
-    .expect("ordinary authorized revocation of one component");
-    assert!(!kernel.storage_mounts.contains_key(&revoked_mount_id));
-
-    assert_replacement_after_unhealthy_hit_for_fresh_execution(&kernel, &caller, &broker, stale)
-        .await;
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn expired_component_invalidates_cached_exact_set() {
-    let (_temporary, kernel, caller, broker) =
-        provider_fixture!("expired_component_invalidates_cached_exact_set");
-    let stale = successful_fleet_mount_for_fresh_execution(&kernel, &caller, &broker).await;
-    for mount_id in &stale.projection.component_mount_ids {
-        kernel
-            .storage_mounts
-            .get(mount_id)
-            .expect("recorded exact component")
-            .expires_at_epoch_secs
-            .store(0, Ordering::Release);
-    }
-
-    assert_replacement_after_unhealthy_hit_for_fresh_execution(&kernel, &caller, &broker, stale)
-        .await;
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn revocation_cannot_interleave_validation_and_cached_reference() {
-    let (_temporary, kernel, caller, broker) =
-        provider_fixture!("revocation_cannot_interleave_validation_and_cached_reference");
-    let stale = successful_fleet_mount_for_fresh_execution(&kernel, &caller, &broker).await;
-    let owner_mount_id = stale.projection.component_mount_ids[1];
-    let owner_state = kernel
-        .storage_mounts
-        .get(&owner_mount_id)
-        .map(|entry| Arc::clone(entry.value()))
-        .expect("owner component");
-
-    let gate = arm_retain_validation_gate();
-    let mount_broker = broker.clone();
-    let mount_caller = caller.clone();
-    let mount_task = Arc::new(OwnedTask::spawn(async move {
-        super::super::process_stop::cleanup_evidence::scoped_with_label(
-            "stale-invalidation-exact-scope",
-            async move {
-                let replacement_mount = super::PROCESS_MOUNT_TEST_ID
-                    .scope(
-                        super::super::process_stop::cache_test_support::fresh_process_mount_test_id(
-                        ),
-                        mount_broker.mount(&mount_caller),
-                    )
-                    .await
-                    .expect("validated unhealthy mount replacement");
-                let evidence_scope =
-                    super::super::process_stop::cleanup_evidence::current_scope_for_test()
-                        .expect("stale invalidation evidence scope");
-                (evidence_scope, replacement_mount)
-            },
-        )
-        .await
-    }));
-    bounded_phase("validation gate entry", gate.entered().notified()).await;
-
-    let revoke_kernel = Arc::clone(&kernel);
-    let revoke_caller = caller.clone();
-    let revocation = Arc::new(OwnedTask::spawn(async move {
-        revoke_lease(
-            &revoke_kernel,
-            &revoke_caller,
-            MountOwnerScope::CallerOnly,
-            owner_mount_id,
-        )
-        .await
-    }));
-    bounded_until("revocation publishes its fence", || {
-        owner_state.revoked.load(Ordering::Acquire)
-    })
-    .await;
-    assert!(
-        bounded_phase(
-            "revocation listener drain",
-            owner_state.wait_listener_closed(),
-        )
-        .await,
-        "revocation must drain the component listener while reference admission is paused"
-    );
-    gate.release().notify_one();
-
-    let body_kernel = Arc::clone(&kernel);
-    let body_broker = broker.clone();
-    let stale_component_mount_ids = stale.projection.component_mount_ids.clone();
-    let finishers = owned_finishers![mount_task, revocation];
-    run_owned_test_body(&finishers, move || async move {
-        let (stale_evidence_scope, replacement_mount) =
-            bounded_phase("validated unhealthy mount replacement", mount_task.join()).await;
-        join_revocation(&revocation, "authorized component revocation").await;
-        assert_ne!(replacement_mount.workspace_root, stale.mount.workspace_root);
-        assert_eq!(
-            stale.projection.refs.load(Ordering::Acquire),
-            1,
-            "the stale projection must not gain a reference after revocation"
-        );
-        assert_eq!(
-            super::super::process_stop::cleanup_evidence::take_for_test(stale_evidence_scope),
-            expected_successful_projection_evidence(&stale_component_mount_ids),
-            "stale invalidation evidence must be consumed before replacement cleanup"
-        );
-
-        close_mount("replacement provider close", replacement_mount).await;
-        close_mount("stale provider close", stale.mount).await;
-        assert!(body_kernel.storage_mounts.is_empty());
-        assert!(body_broker.projections.lock().await.is_empty());
-    })
-    .await;
-}
-
-#[tokio::test]
-async fn stopped_retry_evidence_never_synthesizes_an_acknowledgement() {
-    let fixture = super::exact_fence_fixture().await;
-    let branch_mount_id = fixture.branch.mount_id;
-    let owner_mount_id = fixture.owner.mount_id;
-    let targets = &fixture.binding.targets;
-    let absent_control = fixture.kernel.astrid_home.run_dir().join("absent.sock");
-    let running = |mount_id: StorageMountId| ProjectionLeaseProvider {
-        running: RunningProvider {
-            child: None,
-            control_path: absent_control.clone(),
-            token: "retry-token".to_owned(),
-            stopped: true,
-        },
-        lease: super::ProjectionLeaseTarget {
-            mount_id,
-            target: targets.workspace.clone(),
-        },
-    };
-    let key = ProcessProjectionKey {
-        binding: fixture.binding.clone(),
-        read_write: true,
-    };
-    let state = ProjectionCleanupState {
-        kernel: Arc::downgrade(&fixture.kernel),
-        stop_policy: crate::storage_mount::process_broker::ProcessStopPolicy::default(),
-        binding: fixture.binding.clone(),
-        branch: running(branch_mount_id),
-        owner: running(owner_mount_id),
-        shared: None,
-        mount_root: fixture.kernel.astrid_home.run_dir(),
-        cleaned: false,
-    };
-    let mut projections = std::collections::BTreeMap::new();
-    super::retain_failed_launch_projection(
-        &mut projections,
-        &key,
-        fixture.kernel.astrid_home.run_dir().join("workspace"),
-        fixture.kernel.astrid_home.run_dir().join("owner"),
-        None,
-        state,
-    );
-    let projection = Arc::clone(projections.get(&key).expect("retained retry blocker"));
-
-    let scope = super::super::process_stop::cleanup_evidence::scoped_with_label(
-        "stopped-retry-evidence",
-        async {
-            assert!(
-                !super::retry_failed_projection(&projection, &mut projections, &key).await,
-                "absent production lease state must keep the retry blocked"
-            );
-            super::super::process_stop::cleanup_evidence::current_scope_for_test()
-                .expect("retry evidence scope")
-        },
-    )
-    .await;
-    let events = super::super::process_stop::cleanup_evidence::take_for_test(scope);
-    assert_eq!(
-        events,
-        vec![
-            cleanup_event(ProjectionCleanupStage::Binding, false),
-            provider_stop_outcome_event(
-                ProviderComponent::Branch,
-                branch_mount_id,
-                ProcessStopOutcome::Stopped {
-                    acknowledged: false
-                },
-            ),
-            provider_stop_outcome_event(
-                ProviderComponent::OwnerHome,
-                owner_mount_id,
-                ProcessStopOutcome::Stopped {
-                    acknowledged: false
-                },
-            ),
-        ]
-    );
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn expiry_cannot_interleave_validation_and_cached_reference() {
-    let (_temporary, kernel, caller, broker) =
-        provider_fixture!("expiry_cannot_interleave_validation_and_cached_reference");
-    let stale = successful_fleet_mount_for_fresh_execution(&kernel, &caller, &broker).await;
-
-    let gate = arm_retain_validation_gate();
-    let mount_broker = broker.clone();
-    let mount_caller = caller.clone();
-    let mount_task = Arc::new(OwnedTask::spawn(async move {
-        super::PROCESS_MOUNT_TEST_ID
-            .scope(
-                super::super::process_stop::cache_test_support::fresh_process_mount_test_id(),
-                mount_broker.mount(&mount_caller),
-            )
-            .await
-    }));
-    bounded_phase("expiry validation gate entry", gate.entered().notified()).await;
-    for mount_id in &stale.projection.component_mount_ids {
-        kernel
-            .storage_mounts
-            .get(mount_id)
-            .expect("recorded exact component")
-            .expires_at_epoch_secs
-            .store(0, Ordering::Release);
-    }
-    gate.release().notify_one();
-
-    let body_kernel = Arc::clone(&kernel);
-    let body_broker = broker.clone();
-    let body_caller = caller.clone();
-    let finishers = owned_finishers![mount_task];
-    run_owned_test_body(&finishers, move || async move {
-        let replacement_mount = join_mount(&mount_task, "validated expiry mount replacement").await;
-        assert_replacement_after_unhealthy_hit_for_fresh_execution(
-            &body_kernel,
-            &body_caller,
-            &body_broker,
-            stale,
-        )
-        .await;
-        close_mount("validated expiry provider close", replacement_mount).await;
-        assert!(body_kernel.storage_mounts.is_empty());
-        assert!(body_broker.projections.lock().await.is_empty());
-    })
-    .await;
-}
-
-#[tokio::test]
-async fn externally_removed_member_fences_remaining_exact_authority() {
-    let fixture = super::exact_fence_fixture().await;
-    let removed_mount_id = fixture.shared.mount_id;
-    assert!(
-        fixture
-            .kernel
-            .storage_mounts
-            .remove(&removed_mount_id)
-            .is_some()
-    );
-
-    let cleanup: ProjectionCleanup = Arc::new(|| Box::pin(async { true }));
-    let (projection, key, mut projections) = exact_cached_projection(
-        &fixture,
-        vec![
-            fixture.branch.mount_id,
-            fixture.owner.mount_id,
-            removed_mount_id,
-        ],
-        cleanup,
-    );
-
-    assert!(
-        invalidate_unhealthy_projection(&fixture.kernel, &projection, &mut projections, &key).await
-    );
-    assert!(projections.is_empty());
-    assert_eq!(
-        projection.refs.load(std::sync::atomic::Ordering::Acquire),
-        1,
-        "invalidation must not add a reference to the degraded set"
-    );
-    assert!(
-        fixture
-            .states
-            .iter()
-            .take(2)
-            .all(|state| state.is_revoked_for_test()),
-        "remaining authorized members must be fenced before replacement"
-    );
-}
-
-#[tokio::test]
-async fn invalidation_records_cache_removal_and_complete_success() {
-    let fixture = super::exact_fence_fixture().await;
-    let stale_cleanup: ProjectionCleanup = Arc::new(|| Box::pin(async { true }));
-    let ids = vec![
-        fixture.branch.mount_id,
-        fixture.owner.mount_id,
-        fixture.shared.mount_id,
-    ];
-    let (projection, key, _) = exact_cached_projection(&fixture, ids.clone(), stale_cleanup);
-    let replacement_cleanup: ProjectionCleanup = Arc::new(|| Box::pin(async { true }));
-    let (_, _, mut projections) = exact_cached_projection(&fixture, ids, replacement_cleanup);
-
-    let scope = super::super::process_stop::cleanup_evidence::scoped_with_label(
-        "invalidation-cache-removal",
-        async {
-            let invalidated = invalidate_unhealthy_projection(
-                &fixture.kernel,
-                &projection,
-                &mut projections,
-                &key,
-            )
-            .await;
-            let scope = super::super::process_stop::cleanup_evidence::current_scope_for_test()
-                .expect("typed evidence execution");
-            (scope, invalidated)
-        },
-    )
-    .await;
-    assert!(scope.1);
-    let events = super::super::process_stop::cleanup_evidence::take_for_test(scope.0);
-    assert_eq!(events.len(), 3);
-    assert!(!events.iter().any(|event| event.failed));
-    assert_eq!(events[1].stage, ProjectionCleanupStage::CacheRemoval);
-    assert_eq!(events[2].stage, ProjectionCleanupStage::Complete);
-}
-
-#[tokio::test]
-async fn cleanup_evidence_names_first_lease_resource_failure() {
-    let fixture = super::exact_fence_fixture().await;
-    let owner_mount_id = fixture.owner.mount_id;
-    let owner_state = Arc::clone(&fixture.states[1]);
-    crate::storage_mount::inject_cleanup_fault_for_test(
-        &owner_state,
-        crate::storage_mount::MountCleanupStage::Callback,
-    );
-    let targets = &fixture.binding.targets;
-    let branch = lease_target(fixture.branch.mount_id, targets.workspace.clone());
-    let owner = lease_target(owner_mount_id, targets.owner_home.clone());
-
-    let scope = super::super::process_stop::cleanup_evidence::scoped_with_label(
-        "first-failure-evidence",
-        async {
-            assert!(
-                !revoke_projection_leases(&fixture.kernel, &fixture.binding, &branch, &owner, None)
-                    .await
-            );
-            super::super::process_stop::cleanup_evidence::current_scope_for_test()
-                .expect("typed evidence execution")
-        },
-    )
-    .await;
-    let events = super::super::process_stop::cleanup_evidence::take_for_test(scope);
-    assert_eq!(
-        events,
-        vec![
-            cleanup_event(ProjectionCleanupStage::ListenerSettlement, false),
-            resource_event(fixture.branch.mount_id, false),
-            resource_event(owner_mount_id, true),
-        ]
-    );
 }
 
 #[tokio::test]
